requirements.txt
cycler==0.10.0
imbalanced-learn==0.6.2
joblib==0.14.1
kiwisolver==1.1.0
matplotlib==3.1.2
numpy==1.18.1
pandas==0.25.3
#pkg-resources==0.0.0
pyparsing==2.4.6
python-dateutil==2.8.1
pytz==2019.3
scikit-learn==0.22.2.post1
scipy==1.4.1
six==1.14.0
sklearn==0.0
sentence-transformers<=0.2.6.1
sent2vec-prebuilt==0.0.3
mpi4py==3.0.3


